requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
filelock==3.4.2
Flask==2.0.2
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.22.0
packaging==21.3
pyparsing==3.0.6
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
sentencepiece==0.1.96
six==1.16.0
smmap==5.0.0
tokenizers==0.8.0rc4
tqdm==4.62.3
transformers==3.0.0
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
